+++ src/memory.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
 
-pub fn growCapacity(capacity: u64) u64 {
+pub fn growCapacity(capacity: usize) usize {
     if (capacity < 8) {
         return 8;
     }
@@ -8,6 +8,6 @@ pub fn growCapacity(capacity: u64) u64 {
     return capacity * 2;
 }
 
-pub fn reallocateArray(comptime T: type, ptr: []T, new_size: u64, alloc: std.mem.Allocator) ![]T {
+pub fn reallocateArray(comptime T: type, ptr: []T, new_size: usize, alloc: std.mem.Allocator) ![]T {
     return alloc.realloc(ptr, new_size) catch |err| (return err);
 }
